Revert "extract seedHelpers into shared seeds.go for cross-test use"
This reverts commit b218c77a6eabb2a882c415e525aed864369eefb0.
